==> notebooks/0-data-retrieval.ipynb <==
import os
import pandas as pd
import requests
from tqdm.notebook import trange, tqdm

pd.set_option('display.max_columns', 100)

# EBI MGnify catalogs

def select_representatives(df):
    representatives = set(df['Species_rep'].unique())
    return df.loc[df['Genome'].isin(representatives)]

## Human oral catalog

path_ebi_oral = '../raw_data/EBI/human-oral-v1-0'

!mkdir -p $path_ebi_oral/fna
!wget -P files/input/mgnify_meta http://ftp.ebi.ac.uk/pub/databases/metagenomics/mgnify_genomes/human-oral/v1.0/genomes-all_metadata.tsv
!mv 'files/input/mgnify_meta/genomes-all_metadata.tsv' 'files/input/mgnify_meta/genomes-all_metadata_oral.tsv'

df_ebi_oral = select_representatives(pd.read_csv('files/input/mgnify_meta/genomes-all_metadata_oral.tsv', sep='\t'))

df_ebi_oral.to_csv('files/output/mgnify_meta/representative-all_metadata_oral.tsv', index=False, sep='\t')

for mag in df_ebi_oral['Genome']:
    url = f"http://ftp.ebi.ac.uk/pub/databases/metagenomics/mgnify_genomes/human-oral/v1.0/species_catalogue/{mag[:-2]}/{mag}/genome/{mag}.fna"
    !wget -P "$path_ebi_oral/fna" "$url"

## Human gut catalog

path_ebi_gut = '../raw_data/EBI/human-gut-v2-0-1'

!mkdir -p $path_ebi_gut/fna
!wget -P files/input/mgnify_meta http://ftp.ebi.ac.uk/pub/databases/metagenomics/mgnify_genomes/human-gut/v2.0.1/genomes-all_metadata.tsv 
!mv files/input/mgnify_meta/genomes-all_metadata.tsv files/input/mgnify_meta/genomes-all_metadata_gut.tsv

df_ebi_gut = select_representatives(pd.read_csv('files/input/mgnify_meta/genomes-all_metadata_gut.tsv', sep='\t'))

df_ebi_gut.to_csv('files/output/mgnify_meta/representative-all_metadata_gut.tsv', index=False, sep='\t')

for mag in df_ebi_gut['Genome']:
    url = f"http://ftp.ebi.ac.uk/pub/databases/metagenomics/mgnify_genomes/human-gut/v2.0.1/species_catalogue/{mag[:-2]}/{mag}/genome/{mag}.fna"
    !wget -P "$path_ebi_gut/fna" "$url"

# JGI GEM

path_jgi = '../raw_data/JGI'

!mkdir -p $path_jgi/fna
!wget -P $path_jgi https://portal.nersc.gov/GEM/genomes/genome_metadata.tsv 

df_jgi = pd.read_csv(os.path.join(path_jgi, 'genome_metadata.tsv'), sep='\t')

mask = df_jgi['ecosystem_category'] == 'Human'
df_jgi.loc[mask].to_csv('files/input/jgi_meta/genome_metadata_human.tsv', sep='\t', index=False)

for mag in df_jgi['genome_id']:
    url = f'https://portal.nersc.gov/GEM/genomes/fna/{mag}.fna.gz'
    !wget -P "$path_jgi/fna" "$url" && gunzip $path_jgi/fna/$(basename $url)

# HMP

path_hmp = '../raw_data/HMP'

!mkdir -p $path_hmp/dwnl

## Retrieving files from the wgs_assembled_seq_set category

### Metadata

url = "https://portal.hmpdacc.org/api/files"

params = {
    "fields": "file_format,file_type,file_annotation_pipeline,file_matrix_type",
    "filters": '{"op":"and","content":[{"op":"in","content":{"field":"file.format","value":["FASTA"]}},{"op":"in","content":{"field":"file.node_type","value":["wgs_assembled_seq_set"]}}]}',
    "from": 0,
    "save": "",
    "size": "100",
    "sort": "file_id:asc",
}

all_dfs = []
for params['from'] in tqdm(range(0, 4000, 100)): # <--- increase the range for next pages
    data = requests.get(url, params=params).json()
    all_dfs.append(pd.DataFrame([h['file'] for h in data['data']['hits']]))

df = pd.concat(all_dfs).reset_index(drop=True)

### Deduplication

Keep only the latest assembly for a sample

df['basename'] = df['file_name'].map(os.path.basename)

df.loc[df['basename'].str.contains('SRS011061')]

mask = ~df['srs'].isna()
df_derep = df.loc[mask].sort_values(['srs', 'version'], ascending=False).groupby('srs').head(1)
df_derep = pd.concat([df_derep, df.loc[~mask]])

df_derep.to_csv('files/input/hmp_meta/wgs_assembled_seq_set.tsv', sep='\t', index=False)

### Download links

# list of files to download
df_derep['basename'] = df_derep['file_name'].map(os.path.basename)
df_derep['file_name'].to_csv(os.path.join(path_hmp, 'links_latest_wgs_assembled_seq_set.txt'), index=False, header=False, sep='\t')
df_derep[['md5', 'basename']].to_csv(os.path.join(path_hmp, 'links_latest_wgs_assembled_seq_set.md5'), index=False, header=False, sep=' ')

The files were retrieved with `cd $path_hmp/dwnl && parallel -j 20 -a ../files_wgs_assembled_seq_set.txt  wget {}` and unpacked to `$path_hmp/fna` under the names following `{id}.fna` pattern.
